Changing keyword options and method name.
