--- lib/index.js
@@ -1,10 +1,0 @@
-
-/*!
- * Analytics Machine - Plugins
- * Copyright(c) Thomas Blobaum
- * MIT Licensed
- */
-
-require('directory')(function (fn, filename) {
-  module.exports[filename] = fn
-})
